Restructured dir and test dir
- Renamed widgets to shared and add directories to test dir.
