+++ test/screen/find_pw_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:bookoo2/screens/find_pw.dart' show FindPw;
-import './test_utils.dart' show TestUtils;
+import '../test_utils.dart' show TestUtils;
 
 void main() {
   testWidgets("Widget", (WidgetTester tester) async{
