+++ test/screen/splash_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:bookoo2/screens/splash.dart';
-import './test_utils.dart' show TestUtils;
+import '../test_utils.dart' show TestUtils;
 
 void main() {
   Future future;
